sqlite_to_postgres/requirements.txt
pytest==7.2.1
psycopg2==2.9.5
python-dotenv==0.21.1
